requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.1
langchain>=0.2.3
langchain-community>=0.2.4
langchain-openai>=0.1.7
openai>=1.40.0
faiss-cpu>=1.8.0
pypdf>=4.2.0
docx2txt>=0.8
sentence-transformers>=3.0.0
langchain-huggingface>=0.0.7
